+++ source_gen/source_gen/lib/source_gen.dart
@@ -5,6 +5,7 @@
 library source_gen;
 
 export 'src/builder.dart';
+export 'src/find_type.dart' show findType;
 export 'src/generator.dart';
 export 'src/generator_for_annotation.dart';
 export 'src/type_checker.dart' show TypeChecker;